+++ tests/conftest.py
@@ -1,23 +1,13 @@
-
 import pytest
 from app import create_app
 
 
-@pytest.fixture
-def app():
-
-    app = create_app({
-        'TESTING': True,
-    })
-
-    yield app
-
-
-@pytest.fixture
-def client(app):
-    return app.test_client()
-
-
-@pytest.fixture
-def runner(app):
-    return app.test_cli_runner()
+@pytest.fixture(scope="class")
+def client(request):
+    the_app = create_app(
+        {
+            "TESTING": True,
+        }
+    )
+    client = the_app.test_client()
+    request.cls.client = client