+++ tests/test_dalton.py
@@ -1,5 +1,10 @@
+import unittest
 
-def test_dalton_about(client):
-    response = client.get('/dalton/about')
-    assert b'About Dalton' in response.data
+import pytest
 
+
+@pytest.mark.usefixtures("client")
+class TestDalton(unittest.TestCase):
+    def test_dalton_about(self):
+        response = self.client.get("/dalton/about")
+        assert b"About Dalton" in response.data